move commen functions to utils.py
